+++ docs/CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 0.1.0
 ------
 
-The first release of GeoPySpark! After being in development for the past 5
+The first release of GeoPySpark! After being in development for the past 6
 months, it is now ready for its initial release! Since nothing has been changed
 or updated per se, we'll just go over the features that will be present in
 0.1.0.
@@ -24,10 +24,20 @@ or updated per se, we'll just go over the features that will be present in
  - Perform local and focal operations on ``TiledRasterRDD``.
  - Read, write, and query GeoTrellis tile layers.
  - Read tiles from a layer.
+ - Added ``PngRDD`` to make rendering to PNGs more efficient.
+ - Added ``RDDWrapper`` to provide more functionality to the RDD classes.
+ - Polygonal summary methods are now available to ``TiledRasterRDD``.
+ - Eublidean distance added to ``TiledRasterRDD``.
+ - Neighborhoods submodule added to make focal operations easier.
+
+**geopyspark.command**
+
+ - GeoPySpark can now be used a script to download the jar.
+   Used when installing GeoPySpark from pip.
 
 **Documentation**
 
  - Added docstrings to all python classes, methods, etc.
- - Core-Concepts.
+ - Core-Concepts, rdd, geopycontext, and catalog.
  - Ingesting and creating a tile server with a greyscale data.
  - Ingesting and creating a tile server with data from Sentinel.